Remove code dublication in fused gelu_bw compute kernel
